telegram-bot-service/requirements.txt
python-telegram-bot[socks]==20.7
httpx[socks]==0.25.2
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
